requirements.txt
pandas
requests
python-telegram-bot==13.15
